pyproject.toml
[project]
name = "logosphere"
version = "0.1.0"
description = "Memetic dynamics experiment framework"
requires-python = ">=3.11"
dependencies = [
    "pytest>=9.0.2",
    "requests>=2.31.0",
    "textual>=7.1.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
analysis = [
    "numpy>=1.26.0",
    "scikit-learn>=1.4.0",
    "matplotlib>=3.8.0",
    "hdbscan>=0.8.33",
]
dev = [
    "pytest>=8.0.0",
]

[project.scripts]
logosphere-run = "scripts.run:main"
logosphere-analyze = "scripts.analyze:main"
